Cleaned up errors and warning for 64-bit compile
64-bit compile was tested on 64-bit debian Intel.
This may not be sufficient for 64-bit ARM compilation.

All unit tests pass  on Intel 64-bits.

Also:

- Added guard to call on `invoke()` (vc4), so that it only runs on ARM 32-bits.
- Added header files with #define's for detecting platform dependent stuff
- Some refactoring where appropriate, to get rid of 32/64 bit issues
- Converted tabs to 2 spaces in all files that were touched

+++ Lib/Support/HeapManager.h
@@ -12,43 +12,43 @@ namespace V3DLib {
  */
 class HeapManager {
 public:
-	HeapManager();
-	HeapManager(HeapManager *object) = delete;
+  HeapManager();
+  HeapManager(HeapManager *object) = delete;
 
-	uint32_t size() const { return m_size; }
-	bool empty() const { return m_offset == 0; }
+  uint32_t size() const { return m_size; }
+  bool empty() const { return m_offset == 0; }
 
-	// Intended for unit tests
-	int num_free_ranges() const { return m_free_ranges.size(); }
+  // Intended for unit tests
+  uint32_t num_free_ranges() const { return (uint32_t) m_free_ranges.size(); }
 
 protected:
-	int alloc_array(uint32_t size_in_bytes);
-	void dealloc_array(uint32_t index, uint32_t size);
-	void set_size(uint32_t val);
-	void clear();
-	bool is_cleared() const;
+  int alloc_array(uint32_t size_in_bytes);
+  void dealloc_array(uint32_t index, uint32_t size);
+  void set_size(uint32_t val);
+  void clear();
+  bool is_cleared() const;
 
 private:
-	// Disallow assignment
-	void operator=(HeapManager a);
-	void operator=(HeapManager& a);
+  // Disallow assignment
+  void operator=(HeapManager a);
+  void operator=(HeapManager& a);
 
   uint32_t m_size   = 0;  // Total allocated size of derived heap/buffer object
-	uint32_t m_offset = 0;
+  uint32_t m_offset = 0;
 
-	struct FreeRange {
-		uint32_t left  = 0;
-		uint32_t right = 0;
+  struct FreeRange {
+    uint32_t left  = 0;
+    uint32_t right = 0;
 
-		FreeRange(uint32_t l, uint32_t r) : left(l), right(r) {}  // required by std::vector
-		unsigned size() const;
-		bool empty() const { return size() == 0; }
-	};
+    FreeRange(uint32_t l, uint32_t r) : left(l), right(r) {}  // required by std::vector
+    unsigned size() const;
+    bool empty() const { return size() == 0; }
+  };
 
-	std::vector<FreeRange> m_free_ranges;
+  std::vector<FreeRange> m_free_ranges;
 
-	bool check_available(uint32_t n);
-	int alloc_intern(uint32_t size_in_bytes);
+  bool check_available(uint32_t n);
+  int alloc_intern(uint32_t size_in_bytes);
 };
 
 }  // namespace V3DLib
